refactor: make slideController file to include functions as to slide
