+++ public/js/main.js
@@ -1,12 +1,10 @@
 import * as common from "./common.js";
+import * as slide from "./slideController.js";
 import {weeklyPublicationTemplate} from "./weeklyPublication.js";
 import {homeTemplate} from "./home.js";
 const weeklyWebtoonList = await common.fetchData("weeklyWebtoonList");
 let previousSubCategory = document.createElement("div");
 let previousWeekdayCategory = document.createElement("div");
-let webtoonSlideIndex = 0;
-let isMovable = true;
-let isAutoSlideScheduled = null;
 
 init();
 
@@ -22,13 +20,13 @@ function renderSubCategory(subCategory) {
     case "요일연재":
       subCategorySection.innerHTML = weeklyPublicationTemplate;
       attachWebtoonBoardHandler();
-      attachSlideButtonHandler();
-      executeAutoSlide();
+      slide.attachSlideButtonHandler();
+      slide.executeAutoSlide();
       break;
     case "홈":
       subCategorySection.innerHTML = homeTemplate;
-      attachSlideButtonHandler();
-      executeAutoSlide();
+      slide.attachSlideButtonHandler();
+      slide.executeAutoSlide();
       break;
     case "웹툰": case "소년": case "드라마": case "로맨스": case "로판": case "액션무협": case "BL":
       subCategorySection.innerHTML = "";
@@ -52,23 +50,6 @@ function attachWebtoonBoardHandler() {
   common.delegateEventListener(weekdayCategoryNav, clickWeekdayCategoryNav);
 }
 
-function attachSlideButtonHandler() {
-  const slideButtonDiv = document.querySelector(".slide_button");
-  slideButtonDiv.addEventListener("click", ({target}) => {
-    isMovable = false;
-    if (target.closest(".next_button")) {
-      webtoonSlideIndex++;
-      moveWebtoonSlide();
-      return;
-    }
-    if (target.closest(".previous_button")) {
-      webtoonSlideIndex--;
-      moveWebtoonSlide();
-      return;
-    }
-  });
-}
-
 function clickSubCategoryNav(subCategoryNav) {
   previousSubCategory.classList.remove("bold");
   subCategoryNav.classList.add("bold");
@@ -81,46 +62,4 @@ function clickWeekdayCategoryNav(weekdayCategoryNav) {
   weekdayCategoryNav.classList.add("paint_yellow");
   previousWeekdayCategory = weekdayCategoryNav;
   renderWebtoonBoard(weekdayCategoryNav.innerText);
-}
-
-function moveWebtoonSlide() {
-  const promotionWebtoonSection = document.querySelector(".promotion_webtoon_section");
-  if (!promotionWebtoonSection) {
-    return; // "웹툰", "소년", "드라마", "로맨스", "로판", "액션무협", "BL"의 promotionWebtoonSection은 구현하지 않았으므로 막아두었습니다.
-  }
-  const slideSize = promotionWebtoonSection.clientWidth;
-  const promotionWebtoonSlide = document.querySelector(".promotion_webtoon_slide");
-  const promotionWebtoons = promotionWebtoonSlide.children;
-  if (webtoonSlideIndex < 0) {
-    webtoonSlideIndex = promotionWebtoons.length;
-    promotionWebtoonSlide.style.transform = `translateX(${- slideSize * (promotionWebtoons.length - 1)}px)`;
-    isMovable = true;
-    return;
-  }
-  if (webtoonSlideIndex >= promotionWebtoons.length) {
-    promotionWebtoonSlide.style.transform = "";
-    webtoonSlideIndex = 0;
-    isMovable = true;
-    return;
-  }
-  promotionWebtoonSlide.style.translate = "transform 1s ease-in-out";
-  promotionWebtoonSlide.style.transform = `translateX(${-slideSize * webtoonSlideIndex}px)`;
-  isMovable = true;
-}
-
-function executeAutoSlide() {
-  const delayTime = 3000;
-  if (!isMovable || isAutoSlideScheduled) {
-    return;
-  }
-  isAutoSlideScheduled = setTimeout(() => {
-    autoWebtoonSlide();
-  }, delayTime);
-}
-
-function autoWebtoonSlide() {
-  webtoonSlideIndex++;
-  moveWebtoonSlide();
-  isAutoSlideScheduled = null;
-  executeAutoSlide();
 }
